+++ CHANGES.rst
@@ -12,7 +12,7 @@ Changelog
   ``cart_url``, ``show_to_cart``, ``show_checkout`` and ``show_currency``.
   [rnix]
 
-- Add browser view rendering a JS snipper for Cart JS translations. Cart
+- Add browser view rendering a JS snippet for Cart JS translations. Cart
   translations are now handled via message catalogs.
   [rnix]